requirements.txt
alembic==1.8.1
email_validator==1.3.0
Flask==2.2.2
Flask_Login==0.6.2
Flask_Mail==0.9.1
Flask_Migrate==4.0.0
Flask_SQLAlchemy
Flask_WTF==1.0.1
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy
timeago==1.0.16
tinify==1.6.0
Werkzeug==2.2.2
WTForms==3.0.1
gunicorn
psycopg2
WTForms_SQLAlchemy==0.3
elasticsearch==7.17
